requirements.txt
# Core scientific stack
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Gradient boosting & distributional forecasting
lightgbm>=4.0.0
ngboost>=0.3.14
xgboost>=1.7.0

# Optimization (stochastic MILP, LP)
pulp>=2.7.0
pyomo>=6.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration & utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0
click>=8.1.0

# Statistical testing
pingouin>=0.5.3
statsmodels>=0.14.0

# Notebook support (optional)
jupyter>=1.0.0
ipykernel>=6.25.0